A whole bunch more menu and functional work
The following things work now that did not before:
- Level navigation (requiring unlocks from the multiworld)
- Shops (and purchasing from them, requiring level completion)
- Armor Up, Shield Up, Progressive Generator, Solar Shields items
- Option "Excess Armor" is now respected
- Chat and multiworld events are now shown at the bottom of both the menu and
  ingame screens
- Mouse wheel support for scrolling menus
- Multiworld events and voiced callouts use priority sound channels that
  won't be overridden by anything in the game

+++ src/apmenu.h
@@ -2,4 +2,4 @@
 bool ap_titleScreen(void);
 bool ap_connectScreen(void);
 
-Uint16 ap_itemScreen(void);
+int ap_itemScreen(void);
